Update backend_conn.js
This file receives the parameter 'json_param', which consists of all the available extracted parameters from Dialogflow API after an initial conversation with the user. To check for available devices, the necessary parameters are passed to backend, along with a retrieved DeviceID. The number of available devices with the associated specifications are finally returned back.

Function for issuing a device is yet to be developed.
Problems: Where to store the retrieved parameters for the second function call while issuing the device.